# src/doc/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "API de Gerenciamento de Fazendas",
      "description": "API para gerenciar produção de leite das fazendas",
      "version": "1.0.0"
    },
    "paths": {
...
      "/users/register": {
        "post": {
          "summary": "Registrar um novo usuário",
          "tags": ["Usuários"],
          "parameters": [
            {
              "in": "body",
              "name": "user",
              "required": true,
              "schema": {
                "$ref": "#/components/schemas/UserInput"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Usuário registrado com sucesso",
              "content": {
                "application/json": {
                  "example": {
                    "message": "Usuário registrado com sucesso",
                    "newUser": {
                      "name": "John Doe",
                      "email": "john@example.com"
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Dados de entrada inválidos"
            },
            "500": {
              "description": "Erro ao registrar o usuário"
            }
          }
        }
      },
      "/users/{id}": {
        "get": {
          "summary": "Consultar informações de um usuário por ID",
          "tags": ["Usuários"],
          "security": [{ "bearerAuth": [] }],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Consulta de informações de usuário realizada com sucesso",
              "content": {
                "application/json": {
                  "example": {
                    "_id": "61597c90ebeedf0019e85734",
                    "name": "John Doe",
                    "email": "john@example.com"
                  }
                }
              }
            },
            "404": {
              "description": "Usuário não encontrado"
            },
            "500": {
              "description": "Erro ao consultar o usuário"
            }
          }
        },
        "put": {
          "summary": "Atualizar informações de um usuário por ID",
          "tags": ["Usuários"],
          "security": [{ "bearerAuth": [] }],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "schema": {
                "type": "string"
              }
            },
            {
              "in": "body",
              "name": "user",
              "required": true,
              "schema": {
                "$ref": "#/components/schemas/UserInputUpdate"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Informações de usuário atualizadas com sucesso",
              "content": {
                "application/json": {
                  "example": {
                    "message": "Usuário atualizado com sucesso",
                    "user": {
                      "_id": "61597c90ebeedf0019e85734",
                      "name": "John Doe",
                      "email": "john@example.com"
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Dados de entrada inválidos"
            },
            "404": {
              "description": "Usuário não encontrado"
            },
            "500": {
              "description": "Erro ao atualizar o usuário"
            }
          }
        },
        "delete": {
          "summary": "Remover um usuário por ID",
          "tags": ["Usuários"],
          "security": [{ "bearerAuth": [] }],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Usuário removido com sucesso",
              "content": {
                "application/json": {
                  "example": {
                    "message": "Usuário removido com sucesso"
                  }
                }
              }
            },
            "404": {
              "description": "Usuário não encontrado"
            },
            "500": {
              "description": "Erro ao remover o usuário"
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
      "UserInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "password": {
            "type": "string",
            "example": "secret"
          }
        },
        "required": ["name", "email", "password"]
      }
      }
    }
  }